git-svn-id: http://svn.seqan.de/seqan/trunk@3740 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/demos/find_motif.cpp
@@ -8,7 +8,7 @@ using namespace seqan;
 template <typename TMotifFinder>
 void printMotifs(TMotifFinder & finder)
 {
-	for (int i = 0; i < motifCount(finder); ++i)
+	for (int i = 0; i < (int) motifCount(finder); ++i)
 	{
 		::std::cout << i << ": " << getMotif(finder, i) << ::std::endl;
 	}
